backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# LangChain & AI
langchain==0.3.10
langchain-community==0.3.10
langchain-openai==0.2.10
langchain-anthropic==0.3.0
langchain-google-genai==2.0.4
openai==1.54.0
anthropic==0.40.0
google-generativeai==0.8.3

# Vector DB
qdrant-client==1.12.1

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
youtube-transcript-api==0.6.2

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0

# File Watching
watchdog==5.0.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
